requirements.txt
pandas
psycopg2-binary
pymongo
fastapi
uvicorn
python-dotenv
pydantic
starlette